Move function to v1beta1helper package
