+++ pkg/apis/core/v1beta1/helper/shoot_test.go
@@ -1541,4 +1541,51 @@ var _ = Describe("Helper", func() {
 		Entry("with AutoInPlaceUpdate update strategy", ptr.To(gardencorev1beta1.AutoInPlaceUpdate), true),
 		Entry("with ManualInPlaceUpdate  update strategy", ptr.To(gardencorev1beta1.ManualInPlaceUpdate), true),
 	)
+
+	Describe("CalculateDataStringForKubeletConfiguration", func() {
+		var kubeletConfig *gardencorev1beta1.KubeletConfig
+
+		BeforeEach(func() {
+			kubeletConfig = &gardencorev1beta1.KubeletConfig{}
+		})
+
+		It("should return an empty string if the kubelet config is nil", func() {
+			Expect(CalculateDataStringForKubeletConfiguration(nil)).To(BeEmpty())
+		})
+
+		It("should return an empty string if the kubelet config is empty", func() {
+			Expect(CalculateDataStringForKubeletConfiguration(kubeletConfig)).To(BeEmpty())
+		})
+
+		It("should return the correct data string for the kubelet config", func() {
+			kubeletConfig = &gardencorev1beta1.KubeletConfig{
+				CPUManagerPolicy: ptr.To("static"),
+				EvictionHard: &gardencorev1beta1.KubeletConfigEviction{
+					ImageFSAvailable:  ptr.To("200Mi"),
+					ImageFSInodesFree: ptr.To("1k"),
+					MemoryAvailable:   ptr.To("200Mi"),
+					NodeFSAvailable:   ptr.To("200Mi"),
+					NodeFSInodesFree:  ptr.To("1k"),
+				},
+				SystemReserved: &gardencorev1beta1.KubeletConfigReserved{
+					CPU:              ptr.To(resource.MustParse("1m")),
+					Memory:           ptr.To(resource.MustParse("1Mi")),
+					PID:              ptr.To(resource.MustParse("1k")),
+					EphemeralStorage: ptr.To(resource.MustParse("100Gi")),
+				},
+				KubeReserved: &gardencorev1beta1.KubeletConfigReserved{
+					CPU:              ptr.To(resource.MustParse("100m")),
+					Memory:           ptr.To(resource.MustParse("2Gi")),
+					PID:              ptr.To(resource.MustParse("15k")),
+					EphemeralStorage: ptr.To(resource.MustParse("42Gi")),
+				},
+			}
+
+			Expect(CalculateDataStringForKubeletConfiguration(kubeletConfig)).To(ConsistOf(
+				"101m-2049Mi-16k-142Gi",
+				"200Mi-1k-200Mi-200Mi-1k",
+				"static",
+			))
+		})
+	})
 })
