+++ pkg/component/extensions/operatingsystemconfig/operatingsystemconfig.go
@@ -1081,7 +1081,7 @@ func KeyV2(
 	}
 
 	if !inPlaceUpdate {
-		data = append(data, components.CalculateDataStringForKubeletConfiguration(kubeletConfiguration)...)
+		data = append(data, v1beta1helper.CalculateDataStringForKubeletConfiguration(kubeletConfiguration)...)
 	}
 
 	var result string
