+++ pkg/component/extensions/operatingsystemconfig/original/components/kubelet.go
@@ -5,14 +5,11 @@
 package components
 
 import (
-	"fmt"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	kubeletconfigv1beta1 "k8s.io/kubelet/config/v1beta1"
 	"k8s.io/utils/ptr"
 
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	v1beta1helper "github.com/gardener/gardener/pkg/apis/core/v1beta1/helper"
 )
 
 // ConfigurableKubeletCLIFlags is the set of configurable kubelet command line parameters.
@@ -212,32 +209,4 @@ func reservedFromKubeletConfig(reserved *gardencorev1beta1.KubeletConfigReserved
 	}
 
 	return out
-}
-
-// CalculateDataStringForKubeletConfiguration returns a data string for the relevant fields of the kubelet configuration.
-func CalculateDataStringForKubeletConfiguration(kubeletConfiguration *gardencorev1beta1.KubeletConfig) []string {
-	data := []string{}
-
-	if kubeletConfiguration == nil {
-		return data
-	}
-
-	if resources := v1beta1helper.SumResourceReservations(kubeletConfiguration.KubeReserved, kubeletConfiguration.SystemReserved); resources != nil {
-		data = append(data, fmt.Sprintf("%s-%s-%s-%s", resources.CPU, resources.Memory, resources.PID, resources.EphemeralStorage))
-	}
-	if eviction := kubeletConfiguration.EvictionHard; eviction != nil {
-		data = append(data, fmt.Sprintf("%s-%s-%s-%s-%s",
-			ptr.Deref(eviction.ImageFSAvailable, ""),
-			ptr.Deref(eviction.ImageFSInodesFree, ""),
-			ptr.Deref(eviction.MemoryAvailable, ""),
-			ptr.Deref(eviction.NodeFSAvailable, ""),
-			ptr.Deref(eviction.NodeFSInodesFree, ""),
-		))
-	}
-
-	if policy := kubeletConfiguration.CPUManagerPolicy; policy != nil {
-		data = append(data, *policy)
-	}
-
-	return data
 }
